// public/app/types/devicemanagement/inventory/inventory.ts
export const inventoriesPageLimit = 50;

export interface Inventory {
  id: number;
  type: string;
  uuid: string;
  resource_name: string;
  resource_org: number;
  assigned: boolean;
}

export interface CreateInventoryDTO {
  type: string;
  uuid: string;
}

export interface UpdateInventoryDTO {
  uuid: string;
  type?: string;
}

export interface InventoriesState {
  inventories: Inventory[];
  inventoriesCount: number;
  searchQuery: string;
  searchPage: number;
  hasFetched: boolean;
}

export interface InventoryState {
  inventory: Inventory;
}